requirements.txt
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.36.0
accelerate>=0.25.0
datasets>=2.16.0
safetensors>=0.4.0
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0
pandas>=2.1.0
pyyaml>=6.0
tqdm>=4.66.0
wandb>=0.16.0
einops>=0.7.0
pillow>=10.0.0
matplotlib>=3.8.0
seaborn>=0.13.0
netcal>=1.3.0
uncertainty-toolbox>=0.1.0
paperbanana>=0.1.0
